ゲーム開始時の処理が正しいタイミングで呼ばれていないバグを修正しました
