+++ BalloonGame/Assets/Prefabs/Ingame/BPMEvent.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b4e61275b4c3fbd478ee5ad78951b0f9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
