+++ BalloonGame/Assets/Scripts/Sound/BGMTrackChanger.cs
@@ -6,34 +6,22 @@ using UnityEngine;
 public class BGMTrackChanger : MonoBehaviour
 {
     [Header("再生したい音源の指定")]
-    [SerializeField] List<SoundSource> _transitionSource; // BGM_A, BGM_B, BGM_C, ...
+    [SerializeField] List<SoundSource> _transitionSource = default!; // BGM_A, BGM_B, BGM_C, ...
 
-    [SerializeField, Required] AudioSource _audioSource;
-
-    [Header("再生したい音源のBPMを指定してください。\nintro,transitionが一致している必要があります")]
-    [SerializeField] float _beatsPerMinute = 102f;
+    [SerializeField, Required] AudioSource _audioSource = default!;
+    [SerializeField, Required] BPMEvent _bpmEvent = default!;
 
     int _currentTransitionIndex;
-
-    float _secondsPerBeat;
 
     private void Awake()
     {
         //初期化
         //最初はIntroのため-1にしておく
         _currentTransitionIndex = -1;
-        CalculateSecondsPerBeat();
-    }
-
-    private void CalculateSecondsPerBeat()
-    {
-        _secondsPerBeat = 60f / _beatsPerMinute;
     }
 
     public async void PlayNextTrack()
     {
-        var token = this.GetCancellationTokenOnDestroy();
-
         //待機より前にインデックスを上げる
         _currentTransitionIndex++;
 
@@ -44,23 +32,16 @@ public class BGMTrackChanger : MonoBehaviour
             return;
         }
 
-        float timeSinceStart = Time.time;
-        float nextBeatTime = Mathf.Floor(timeSinceStart / _secondsPerBeat) * _secondsPerBeat + _secondsPerBeat;
-        
-        //次のビートまで待機
-        while (timeSinceStart < nextBeatTime)
-        {
-            timeSinceStart += Time.deltaTime;
+        //ビートのタイミングまで待機
+        await _bpmEvent.WaitForBeat();
 
-            await UniTask.Yield(token);
-        }
-
+        //現在の再生位置を取得
         int timeSample = _audioSource.timeSamples;
 
         //再生したい音源を指定
         SoundSource currentSoundSource = _transitionSource[_currentTransitionIndex];
 
-        //再生
+        //取得した再生位置から再生
         SoundManager.Instance.PlayBGM(currentSoundSource, timeSample);
     }
 }
