--- BalloonGame/Assets/Scripts/Sound/BeatEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 431687db735fd8849ba6997a41d4caeb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
